Fix two table accesses in Drawable:setPosition()
As we build the lookup-table to convert positions into XY-coordinates
we try to reuse earlier values in that table.  There are two places
where we want to reuse the X-coordinate for the ‘Center’ position but
we fail to actually access that element, instead copying the table of
both coordinates.  This later causes a crash when we call
love.graphics.draw() because we give it that table instead of the
intended integer coordinate.