Improve the clarity in the instruction documentation about opcodes
The document emphasizes the word ‘directly’ to stress that LNVL does
not directly create instructions from content, instead creating
opcodes first that later become instructions.  But the actual markup
of the document does not drive home this point that clearly.  The
important part to emphasize is that LNVL creates opcodes first, then
instructions, and so the markup should reflect that emphasis.