Replace Instruction.getForOpcode() with a constant lookup table
Every time we call Scene.drawCurrentContent(), which happens
repeatedly during the love.draw() routine, we call the getForOpcode()
function to get the LNVL.Instruction object for the current opcode we
are executing.  The mapping between opcodes and instructions never
changes during run-time.  So we can replace the getForOpcode()
function with a constant lookup table.  In theory this should improve
performance but the difference is realistically negligible.  However,
it is still a better approach to take.