Implement and document the ‘no-op’ instruction
LNVL no longer runs in debugging mode if there is not a corresponding
instruction for each and every opcode.  The ‘no-op’ opcode is never
converted into an instruction for efficiency, since it is going to do
nothing anyways.  But the recent sanity check added to debugging mode
still requires an instruction for the opcode, even if we do never
perform the conversion.  This patch adds the ‘no-op’ instruction,
which also does nothing.  We still never actually use it, but we do
add and document it to satisfy the checks in debugging mode.