Remove a slightly misleading reference to ‘HTML’ documentation
A comment for the LNVL.Instructions table references an HTML document
mentioned at the start of the code.  We use that document to generate
an HTML document, but it is not an HTML document itself.  Since
the project does not come with that HTML it is better to not mention
it explicitly, and instead only refer to the text document in the
repository.