Update the constructor calls in the ‘04’ example script
The script continues to use the old constructors, not the new()
methods that have been part of LNVL for quite a while now.  This patch
also changes the calls to the constructor for LNVL.Menu even though
that class does not yet exist.  We will use new() for its constructor
when we do write it, for consistency with the rest of the classes, so
we can go ahead and change the calls in the example.