Create the LNVL.MenuChoice class
It will be useful in the code to have a way to differentiate tables
representing menu choices from other tables.  This class will serve
that purpose.  It will also provide us a way to more conveniently
access data in menu choices via named properties, as opposed to
relying on table indices which could change in the future.