Store choices for a menu when calling the LNVL.Menu constructor
This patch converts the arguments for the constructor into
LNVL.MenuChoice objects.  This code causes the ‘04’ example script to
break since that script uses strings for the so-called ‘action’
parameter of each choice, while the implementation demands that each
‘action’ be a function.  Later we will provide logic that turns those
strings, which name LNVL.Scene objects, into functions that will cause
the flow of a story to jump to those scenes.