Document in code possible return values for ‘MenuChoice.action()’
The initial design of the property—a function—was that it would never
return anything.  However, it is useful in retrospect to allow the
action function for a menu choice to potentially return one or more
LNVL.Opcode objects.  This will make it possible for menus to affect
scenes by returning existing opcodes for the scene to process instead
of having to re-implement the behavior of those opcodes.