Allow scene names as strings for the actions of menu choices
We require the ‘action’ property of a MenuChoice object to be a
function which may return opcodes.  However, for convenience we allow
the values for actions in scripts to be strings that represent the
names of scenes.  The implied action is to switch to the named scene
if the player chooses that item from the menu.  For example, we allow
menu definitions like this in scenes:

    LNVL.Menu:new{
        {"Enter", "ENTER_SCENE"},
        {"Don't Enter", "LEAVE_SCENE"},
    }

LNVL should switch to one of the two scenes, ‘ENTER_SCENE’ or
‘LEAVE_SCENE’, depending on what the player chooses.  In this case we
assume those strings are the names of LNVL.Scene objects that the game
author creates somewhere.

This patch makes the LNVL.Menu constructor aware of these types of
strings.  When it sees a string it will first convert that to an
anonymous function that returns a ‘change-scene’ opcode.  Then the
MenuChoice constructor will not stop the program with an error since
we will always correctly give it a function for the ‘action’
property.

We can now run the ‘04’ example script again with this patch in place,
although the scene changes do not work because we have not made LNVL
aware of the opcodes that menu choices can create.