Create an example script that changes global settings
This script serves as an example and a test of changing global
settings in a dialog script.  Normally we change these settings within
the ‘src/settings.lua’ file.  But because of the way LNVL loads these
scripts it is possible to change some settings in a script itself
before we create any objects that rely on those settings, like
Scene objects.