Load scripts in their own environment
This patch changes LNVL.loadScript() so that it processes scripts in
their own environment.  We want to do this since scripts must have
variable assignments like

    START = LNVL.Scene:new{…}

and we do not want that to clobber any potential ‘START’ variable in a
game using the LNVL engine.  Global variables created in LNVL scripts
should not trample any variables outside of those scripts.  In this
patch we create a separate environment for each script so that we can
put those global assignments into a sandbox.