Use a global table for the LNVL script environment
This patch removes the local environment table from LNVL.loadScript()
and replaces it with the global LNVL.ScriptEnvironment table.  Some
benefits to this approach are:

1. All scripts we load will share the same environment and that will
make it easier for one script to rely on definitions in another.  For
example, we can define all characters for a story in one script and
have access to those characters everywhere else.

2. The LNVL engine has access to the environment for scripts and can
modify it dynamically.  This makes it possible for the LNVL modules to
augment the environment with additional data, functions, etc.

This change does cost us some amount of privacy since we now expose
the environment table, but that is worth the usefulness we get from
this approach moving forward.