Report any errors in dialog scripts when using debugging mode
The LNVL.LoadScript() function ignores any errors in scripts.  That
may be ok for production releases, but when we are debugging games and
LNVL itself we want to see anything that may go wrong in scripts.
This patch changes the function so that it does not run scripts in
protected mode if we enable LNVL’s debugging mode.  If debugging mode
is not enabled then the engine will silently ignore errors in dialog
scripts like before.