Create aliases for the Scene, Character, and Menu constructors
When we create a Scene inside of a dialog script we write:

    FOO_SCENE = LNVL.Scene:new{…}

Not only is this verbose, it exposes the name of a class and its
constructor.  Showing this information runs contrary to our desire for
a domain-specific language for our dialog scripts.  We do not want to
think about implementation details in our story scripts, and we want
to avoid potentially breaking the validity of our scripts whenever we
change an implementation detail such as a constructor name.

This patch provides a utility function that lets us create aliases for
constructors, which we can then use in scripts.  For example, this
lets us write

    FOO_SCENE = Scene{…}

instead of the more verbose example above.  That makes scripts easier
to write and hides more of the implementation details.  The patch
provides aliases for three objects we commonly create in scripts:
Scenes, Characters, and Menus.