Make LNVL.LoadScript() work with scripts that have no ‘START’ scene
The engine expects a scene with the name ‘START’ to exist because that
is the scene we always use first.  However, the first script we load
in a story may not define that scene.  So this patch teaches the
LNVL.LoadScript() to only assign ‘START’ to LNVL.CurrentScene if
‘START’ exists in the script environment, meaning a script defined the
scene somewhere.

The patch also adds an assertion to ensure that ‘START’ is actually an
LNVL.Scene object.