Clarify in the Howto that each story must define the ‘START’ scene
It is likely that authors will spread a large story across multiple
scripts and load them individually, opposed to writing a single,
behemoth script.  The Howto document says that each script must define
the ‘START’ scene but this is no longer true since LNVL.LoadScript()
now accepts scripts that do not define the scene.  Authors must still
define the ‘START’ scene but only need to make sure they do this
within a single story, not within every script like the Howto
incorrectly suggests.  So this patch clarifies the wording in the
Howto regarding that requirement.