Fix crash in ‘set-scene’ instruction when looking for Scene objects
We now store the Scene objects in the LNVL.ScriptEnvironment table.
But the instruction still looks for those objects in the global
environment where we previously kept them.  This patch changes the
instruction so that it looks for scenes in the correct environment,
preventing the crash that happens when it fails to find a scene, which
always happens if we look in the wrong environment.