Allow function aliases in dialog scripts
This patch provides a utility for creating function aliases that we
can use in dialog scripts.  It is similar to how we create aliases for
class constructors.  The only difference is that it makes an alias for
an arbitrary function and not a class method.  So for example if we
add the code

    LNVL.CreateFunctionAlias("changeToScene", LNVL.Scene.changeTo)

then we can use ‘changeToScene()’ in our story scripts instead of the
more verbose ‘LNVL.Scene.changeTo()’.  This is more economical in
terms of key-strokes, without sacrificing readability, and it helps
hide implementation details such as the underlying engine function
that we really call.