Make ‘Color’ an alias for ‘LNVL.Color’ in dialog scripts
The LNVL.Color module provides short-cuts for color names so that we
can write things like LNVL.Color.NavyBlue in scripts.  Now that we run
scripts in their own environment we can take this short-cut a step
further and create an alias ‘Color’ table so that we only need to
write Color.NavyBlue, omitting the ‘LNVL’ prefix.  This patch adds
that alias and updates all example scripts and the Howto document.