Rename ‘LNVL.Scene.Pause’ to simply ‘Pause’
This patch redefines the constant as part of the script environment
instead of creating a global variable in the LNVL table.  And it gives
the constant a shorter name so that we do not have to write the
verbose prefixes in our dialog scripts.  The patch also updates the
uses in the example scripts and the Howto documentation.