Teach Character objects to draw themselves to the screen
This patch implements a simple Character:draw() method that will
render the character’s avatar to the screen.  This does not provide
any functionality we do not already have, but it will help clarify our
intent in other parts of the code by making it more obvious what we
mean to draw.