Keep a list of active characters in scenes and draw them
One problem with LNVL is that we only draw characters on screen when
they are speaking lines.  If you look at visual novels you will see
that they rarely behave this way.  Once a character enters a scene he
or she almost always remains on screen until explicitly dismissed.

This patch is the first step in making LNVL behave that way.  Be begin
by adding an ‘activeCharacters’ property to the Scene class.  It is a
table that contains all of the active characters in the scene, meaning
the characters which should appear on screen every time we render the
contents of the Scene object.  We perform this rendering when we call
the drawEssentialElements() method.

All we do in this patch is define the table of active characters and
loop through it, drawing each character when we draw the scene.
However, we do not actually populate the table with any Character
objects with this code, so it has no noticeable effect.