Populate the active character list in scenes
We have a list of active characters for each Scene object but this
patch implements the logic that actually puts Character objects into
that list.  We want to activate a character, i.e. add it to that list,
based on the current opcode we are looking at for each piece of
content in a scene.  So for example, if the current opcode is ‘say’
and that opcode has an associated Character object in its ‘arguments’
property then we add that Character to the list of active characters.
That way a character becomes active when he or she speaks.

With this patch in place character images will remain on screen even
when a character is not speaking or performing an action, so long as
the character is active.  But a problem with this patch is that we
have no way to deactivate a character.