Implement Character:leavesTheScene()
This method gives us a way to remove characters from scenes so that
LNVL will stop drawing them.  We need this for story scripts so that
we can do things such as have characters exit a scene.

To remove a character we use a new opcode: ‘deactivate-character’.
The name reflects the fact that we remove a character by taking him or
her out of a scene’s ‘activeCharacters’ table.  We configure the
opcode to generate a ‘no-op’ instruction but we actually avoid
executing the instruction.  When we draw the contents of a Scene
object we see if the current opcode is ‘deactivate-character’, and if
so we remove the character from that list and then immediately stop
before we convert that opcode to a ‘no-op’ instruction.  Since we know
that opcode is a no-op we can exit the drawing routine as soon as we
see it.