Replace the ‘draw-character’ opcode with a ‘move-character’ opcode
Scenes in LNVL originally had no information about characters
appearing in the scene.  The ‘draw-character’ opcode was a way to
explicitly tell LNVL when to render a character on screen.  Another
responsibility of the opcode was to update the position of the
character by changing the ‘position’ property of the character’s
current image.

Now that the Scene class keeps a list of active characters and
automatically draws them we no longer need the ‘draw-character’
opcode.  This patch removes all of the code for situations where we
previously created that opcode.  The active character list inside of
Scene objects takes care of drawing characters for us and we can
safely rely on that without having to tell LNVL when to draw
characters via opcodes.

But we do need to maintain the second responsibility of
‘draw-character’: changing a character’s position as needed.  We keep
this behavior by replacing ‘draw-character’ with the new
‘move-character’ opcode.  That opcode updates the position of a given
Character object.  And in order for the position change to have full
effect we teach Character:draw() to propagate that change.  So if we
change the value of ‘Character.position’ that method will assign the
same value to ‘Character.images[Character.currentImage].position’,
guaranteeing the character appears on screen in the correct place.

A side-effect of this change is that we can make the ‘say’ opcode
processor a no-op.  Previously we would have to generate an additional
‘draw-character’ opcode so that speaking characters would appear on
screen.  But that is no longer necessary and thus the ‘say’ opcode has
nothing else to do during processing.