Return Color objects from Color.FromHex()
Currently we only return a table of RGB values from the function.
While this reflects the structure of Color objects it is missing an
important piece of data: the metatable.  Our failure to set the proper
metatable for the return value of Color.FromHex() causes errors
whenever we try to discriminate tables based on their metatable.

This patch fixes the problem by ensuring that the function returns
tables that have LNVL.Color as their metatable, for consistency with
all other colors we use throughout the engine.