Rewrite Graphics.DrawText() to accept an arbitrary table of contents
The previous design of DrawText() required three arguments:

1. A Font object.

2. A Color object.

3. The actual text to draw.

However, we allowed the third argument to also be a table which could
contain strings or Color objects, indicating that DrawText() should
switch to using those Colors as it came across them.  This was a messy
design.

This patch redesigns the function so that it accepts only one
argument: a table.  That table may contain any sequence of the three
elements listed above.  However, we eliminate the possibility for text
to also be a table containing colors.  This is not necessary since
arguments to DrawText() can now have colors anywhere before strings.

With this redesign we need to change how we process the ‘say’
instruction.  It composes the order of arguments for the function and
so we modify it to fit the new DrawText().  Furthermore, within that
instruction we can now call Graphics.DrawText() directly and eliminate
the Scene:drawText() method.  We used the latter to add font
information but this is no longer required, so we can delete the
method entirely.