Introduce the function LNVL.Initialize()
Any game using LNVL only needs to require the ‘LNVL.lua’ file.  That
file loads all of the LNVL sub-modules.  However, it makes an
assumption about the paths to those sub-modules in every require()
call it uses.  Depending on where the LNVL source is in a game those
paths may be invalid.  We need a way to optionally provide a prefix to
every path we require() to give games the freedom to place the LNVL
source code wherever is most convenient.

The new LNVL.Initialize() function accepts an optional prefix to add
to all paths that we require() for LNVL sub-modules.  This change
means that using LNVL is now a two step process.  For example:

    local LNVL = require "LNVL"
    LNVL.Initialize("path.to.LNVL")

The commentary at the beginning of the main ‘LNVL.lua’ documents this
new requirement.