Update the main example code to call LNVL.Initialize()
This also serves as an example of how we do not always need to give
the function an argument.  This works because the LNVL sub-modules we
require already have the correct paths without any prefix, that is,
when we treat LNVL as an independent program and not a component of a
larger game.