Fix a bug loading ‘src/rgb.txt’ in the Color module
Commit 33be1308 uses the new LNVL.PathPrefix value in an attempt to
make sure LNVL uses the right path when opening the ‘rgb.txt’ file.
This fails, however, because LNVL.PathPrefix always ends in a period.
That terminal period makes sense for when we use LNVL.PathPrefix in
require() statements, but not when opening a file because it gives us
a bad path such as ‘LNVL.src/rgb.txt’.  We must be able to choose when
to use a period or a slash after the prefix.

This patch changes the value we assign to LNVL.PathPrefix so that it
does not have a terminal period.  It then changes the uses of the
value so that we add a period or slash as necessary, making it
possible to load ‘rgb.txt’ in the Color module.