Delete all three Tupfiles
This is the first step in transitioning from Tup to GNU Make.  The
reason for the change is that we run into problems when using LNVL as
a sub-module for our game, which also uses Tup.  Building the game
executes each Tupfile in LNVL, which is the correct behavior for Tup,
but that causes errors and failures to create files.

In all honesty this very likely reflects an error and misunderstanding
on my part, and not a problem with Tup, as it has been a pleasure to
use with other projects.  But I need to get things done and cannot
afford to spend much time tracing down mistakes in Tupfiles.  Make may
be old but it’s reliable and will be easier to control for building our
game, combining resources, etc.