Add a Makefile to replace the deleted Tupfiles
This patch adds a GNU Makefile to the top-level directory which
can perform five tasks:

1. Generating HTML documentation from the Markdown sources by using
the now ‘docs/build-html.sh’ shell script.

2. Create a TAGS file for GNU Emacs.

3. Create the ‘src/settings.lua’ file from the example file that the
project provides.

4. Zip all of the necessary source files into a LÖVE archive.

5. Delete all of the files that the tasks above can create in order to
have a clean slate.

The commentary in the Makefile describes the necessary software that
some of its rules depend on and where to find them, along with a
description of the behavior of running ‘make all’, the most common
use-case.