Implement the basics for ‘add-menu’ and ‘show-menu’
We use no-ops for both this opcode and its corresponding instruction.
In all likelihood we will keep the no-op for the opcode.  However, we
provide a stub implementation of ‘show-menu’ because it will need to
invoke a coroutine that we have not yet written.  The documentation in
the ‘Instructions.md’ file explains the purpose and reasoning.  But we
still add this stub to satisfy the debugging code which ensures that
we are not forgetting to implement any opcode or instruction
processors, otherwise we would not be able to run any other example
script until we completely finish the implementation of ‘show-menu’.