Fix a bug when using named colors in dialog scripts
Trying to write something like

    Lobby = Character { name="Lobby", textColor=Color.Black }

would crash the game because LNVL would not recognize ‘Color.Black’.
More specifically, the if-else branch in Graphics.DrawText() would
fail to recognize the color (i.e. a table) as a valid color because
that function checks metatables to determine what a table represents.
This patch makes ‘Color’ the metatable of every named color available
so that DrawText() and similar functions can recognize such tables as
colors by checking for that metatable.