Create ‘add-menu’ opcodes when we have Menu objects in Scenes
This patch teaches the Scene ‘content to opcode’ conversion process to
be aware of Menu objects.  Everything else we currently use as
arguments to the Scene constructor returns either opcodes or tables of
opcodes.  The Menu class is an exception; we create Menu objects as
arguments to a Scene but have no way to turn them into opcodes.

With this patch a Scene will turn each Menu it finds in its arguments
into an appropriate ‘add-menu’ opcode to process as an instruction
later, although we do not yet implement the instruction side.