Store menus by their name in Scene objects
We add a ‘menus’ table to the Scene class that contains a list of all
of the Menu objects in that scene, accessible by their name, i.e. the
‘Menu.name’ property.  When we give other games access to a Scene to
handle a menu choice, for example, this provides a way for that game
to examine menus the player has previously encountered and use that to
influence what content it shows, what effects occur, et cetera.