Correct the documentation about the ‘show-menu’ instruction
It mentions that developers should refer to the ‘settings.lua’ file in
order to configure the handler.  However, the name of the handler does
not belong to the Settings module, which is necessary if the
definition is actually in the ‘settings.lua’ file.  This patch renames
the handler to something within the Settings module.