Define and document an example menu handler in the settings
This provides a ‘Handlers’ section of the settings file and describes
the overall purpose and nature of handlers, what they mean in the
context of LNVL.  The patch also contains a really dumb implementation
of a menu handler for demonstration purposes.  And so that the engine
will not crash when we actually call that handler later.