Implement the ‘show-menu’ instruction
This makes use of the recent menu handler in the ‘show-menu’
instruction.  It causes no bugs or crashes with example scripts, but
that said we have no examples which really stress or test this logic
to find cracks.  So we cannot say that this is a production-ready
implementation of the instruction.  At best it satisfies the minimum
we require of engine instructions.