Document how ‘handlers’ can be plain functions, not always coroutines
We open up the possibility for handlers to be functions because it
will allow us to provide hooks into LNVL for situations where a
coroutine would be unnecessary, situations where the developer would
not want the game logic to continue until the handler finishes,
e.g. drawing a character to screen before presenting his or her
dialog in a scene.  We can do this with coroutines but to do so is to
simply call functions in a complicated, verbose way for no gain.