Create and document Settings.Handlers.Drawable()
We will use this function as a hook to allow games to render images to
screen in a way they find most fitting, if necessary, as the comments
for the code attempts to explain at length.

Note that this patch does not work.  It assigns the handler a value
that does not exist, i.e. ‘LNVL.Drawable.DefaultHandler’.