Make Settings.Handlers.Drawable a table with two hooks
We need to allow handlers for two Drawable methods:

1. draw()
2. setPosition()

The original design accounted only for the first.  This patch
redefines the Drawable handlers as a table of two functions, one for
each method.  It makes appropriate changes to the documentation to
explain the changes to the API and the arguments those functions
receive.

This patch also addresses a problem posed by the previous commit: how
do we assign a default handler when LNVL loads the Settings module so
early?  We cannot assign any functions from the Drawable module
because we do not have access to it when loading Settings, because we
load the former much later.  So this patch documents the approach we
will take to get around that: assigning temporary ‘nil’ values that we
will change to actual default handlers elsewhere, with documentation
pointing out where other developers can find those defaults.