Implement default functions for the Settings.Handlers.Drawable table
We use the existing Drawable:draw() and Drawable:setPosition() methods
as the default implementations for the two functions required to
satisfy the Settings.Handlers.Drawable table.  We slightly rewrite
those methods as slighty-renamed functions so that we can use the
actual method names as ways to invoke the handlers.  This extra level
of indirection makes it possible for games using LNVL to now customize
how they render Drawable objects to screen.