Restrict the list of valid positions to the Positions module
This patch makes the previously publicly accessible
‘Position.ValidPositions’ table into a local table whose scope does
not extend beyond the Positions module.  We want to provide hooks an
openings for developers to modify LNVL as needed without having to dig
into the LNVL code itself.  But we should have the attitude that
things must be private first, only made public and modifiable if
necessary for games using LNVL.  This will help avoid situations where
a game may rely on a feature that we later want to make private but no
longer can because something already relies on it.

TL;DR Encapsulation