Extract the logic for Opcode.__tostring() into a separate function
This takes all of the code for that metatable function and moves it
into a function local to the Opcode module: formatOpcode().  Now
converting an Opcode into a string involves calling that function.
The motivation for this separation is so that I can occasionally (and
probably temporarily) add extra debugging information strictly inside
the tostring() function.