Add the ‘set-position’ instruction to fix the ‘move-character’ opcode
It turns out the ‘move-character’ opcode has not worked properly for a
long time—which unfortunately I failed to notice.

The opcode would change the ‘position’ property of a Character
object but the result is that we would immediately render that
character in the position of the final ‘move-character’ opcode in the
scene.  That is because we process all opcodes before executing any
instructions, and instructions are what render characters.  Therefore
all of the modifications to the ‘position’ property happen before any
instruction execution and thus any rendering.

The seemingly obvious solution is to defer the change of the
‘position’ property until the instruction execution cycle.  This patch
does exactly by introducing the new ‘set-position’ instruction.  We
change the ‘move-character’ opcode so that it informs the instruction
what object it should affect, i.e. which object has the ‘position’
property it should change.

With this patch in place example scripts like ‘07’ work as intended.
Trying to back up through all of ‘07’ does not completely work, as in
returning the character to its original starting position.  But for
now we can accept scenes not being able to perfectly deal with moving
backwards through content that changes character positions.