Document a handler to use for drawing characters
This patch documents what will be a handler for rendering characters,
i.e. a hook into the Character:draw() method.  It describes the
handler function and its argument, and what the engine expects the
function to do.  We also take this time to shuffle around some
documentation about handlers that have ‘nil’ values and why we must
use those to defer the assignment of actual default handlers until
later in the module loading process.