Implement the code for the default character handler
This commit adds the code for the handler that the previous commit
describes.  All we do is take the existing Character:draw() method and
rename it, assign that as the default handler, and then rewrite
Character:draw() to invoke that handler.  This provides the necessary
gap in the system for game developers to hook in their own behavior
and process for rendering characters if they wish.