Refactor the logic for managing a scene’s active character list
This is a step towards breaking down the logic for rendering scene
content into smaller, more flexible pieces.  Our existing approach is
to update the active character list in a scene whenever we convert the
current opcode to an instruction and execute that to (usually) render
something to the screen.  This patch takes all of the logic for
managing the character list and places it within its own method.

With this patch in place we lose some existing functionality because
we do not call the new method from anywhere.  That means we lose all
management of the active character list.  The plan is to call this
method later when we decide the best place after dividing up more of
the existing Scene logic into smaller chunks.