Implement a method for fetching all current instructions in a Scene
Continuing the refactoring of the code for rendering scenes….  Every
tick we create content on screen by converting the scene’s current
opcode into one or more instructions and then executing those
instructions.  This logic was in the drawCurrentContent() method but
rendering the content and converting the opcode into instructions are
separate steps.  This patch removes that conversion process from
drawCurrentContent() and places it within its own method.