Provide a default implementation for Settings.Handlers.Scene
This completes the conversion and break-down of the Scene code to the
point where we can allow other developers to hook in their own draw()
routine for scenes if they so desire.  The default implementation
provides all of the functionality we originally had, while throwing
away a little bit of code along the way thanks to the refactoring from
previous commits.