Do not create an ‘LNVL’ table in the global namespace
This patch modifies the main ‘LNVL.lua’ file so that it creates a
local table private to the module and then returns that at the end.
The semantic difference and improvement is that this no longer adds a
tabled named ‘LNVL’ to the global namespace.  While unlikely to cause
conflicts it is still better to avoid that possibility entirely.