Revert "Do not create an ‘LNVL’ table in the global namespace"
This reverts commit f772a3bfc71bfa622eb5770ebc8481f9654608a4.

We run into LÖVE blue-screens when loading modules via the
LNVL.Initialize() function.  This is because those modules do not see
the LNVL table.  The purpose of the original change was to avoid
naming conflicts when inserting LNVL into games.  However, I have
changed my mind about this and no longer believe the risk of conflict
is that significant.  So for the time being I would prefer the quick
fix of reverting the above commit.