Character: Rename setName() method to displayName()
A method like setName() makes it sound as if we are changing the
character’s name, which is not its intent.  The method changes what
name LNVL displays.  Hence the new name.

Note that this patch still does not provide an implementation, only
documentation for how the method will behave.