Makefile: Do not require Pandoc or Exuberant Ctags
If the developer has these programs then he can create documentation
in HTML form and tags for use with GNU Emacs.  However, we should not
require these programs in order to build LNVL.  This patch changes the
Makefile so that by default it will only create the engine settings
and the archive of the engine itself.

Traq-Ticket: 16