Fix a crash when refreshing characters in a scene
It is possible for Scene:refreshActiveCharacters() to receive an
opcode that has no arguments, and that implies it has no characters.
If the opcode has no arguments then the function will crash when it
attempts to access character data.  This patch adds a test to see if
there are any opcode arguments, and if not we exit the function.

Traq-Ticket: 15