Remove Character:monologue() by teaching say() to accept lists
Because of the __call() metamethod any strings we give directly to
Character objects act as contents for say().  We implemented
monologue() to accept a table of strings so that we could avoid
writing out a character’s name over and over during a conversation.
For this purpose we used a separate monologue() method in dialog
scripts and had an associated ‘monologue’ opcode.

However, that method and opcode are unnecessary if we rewrite the
‘say’ opcode to accept either a string or a table.  If given a string
then ‘say’ behaves the way it always has.  But now if we give it a
table we see results as if we used monologue().

Overloading say() in this way, so that we can remove monologue(),
helps improve the simplicity of dialog scripts just a tiny amount.
This patch modifies the relevant example to demonstrate.