Implement Character:thinks() method and related opcode
This patch gives all Character objects a thinks() method which is
syntactically the same as says(), but is semantically used to
represent character thoughts.  The method creates one or more new
‘think’ opcodes.  Ideally we would use these opcodes with a specific
instruction to control their display, but for now we convert ’think’
opcodes into ’say’ instructions and print their contents in that way.